fix: rewards share, add keys with excess bond
