+++ features/stealing/stealing-report-form/context/use-stealing-report-submit.ts
@@ -28,7 +28,7 @@ export const useStealingReportSubmit = ({
       nodeOperatorId,
       blockhash,
     }: StealingReportFormInputType): Promise<boolean> => {
-      invariant(amount, 'Amount is not defined');
+      invariant(amount !== undefined, 'Amount is not defined');
       invariant(nodeOperatorId !== undefined, 'NodeOperatorId is not defined');
       invariant(isHex(blockhash), 'BlockHash is not valid');
