+++ package.json
@@ -25,7 +25,7 @@
     "@lidofinance/api-rpc": "^0.47.0",
     "@lidofinance/eth-api-providers": "^0.47.0",
     "@lidofinance/eth-providers": "^0.47.0",
-    "@lidofinance/lido-csm-sdk": "npm:@exromany/lido-csm-sdk@0.1.6",
+    "@lidofinance/lido-csm-sdk": "npm:@exromany/lido-csm-sdk@0.1.7",
     "@lidofinance/lido-ethereum-sdk": "^4.4.0",
     "@lidofinance/lido-ui": "^3.26.0",
     "@lidofinance/next-api-wrapper": "^0.47.0",
