+++ shared/components/key-strikes/index.ts
@@ -1,0 +1,1 @@
+export * from './key-strikes';
